backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
environs==9.5.0
exceptiongroup==1.2.2
filelock==3.15.4
flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
fsspec==2024.6.1
gensim==4.3.3
grpcio==1.63.0
h11==0.14.0
haystack-ai==2.3.0
haystack-experimental==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
lazy-imports==0.3.1
MarkupSafe==2.1.5
marshmallow==3.21.3
milvus-haystack==0.0.9
milvus-lite==2.4.8
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.36.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
posthog==3.5.0
protobuf==5.27.2
pydantic==2.8.2
pydantic-core==2.20.1
pymilvus==2.4.4
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.42.4
typing-extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
